+++ Packages/XCharts/Runtime/Component/Axis/AxisHelper.cs
@@ -354,8 +354,8 @@ namespace XCharts.Runtime
                 axis.splitNumber = splitNumber;
                 return;
             }
-            if (axis.type == Axis.AxisType.Time) { }
-            else if (axis.minMaxType == Axis.AxisMinMaxType.Custom)
+            if (ceilRate == 0) ceilRate = axis.ceilRate;
+            if (axis.minMaxType == Axis.AxisMinMaxType.Custom)
             {
                 if (axis.min != 0 || axis.max != 0)
                 {
@@ -371,13 +371,19 @@ namespace XCharts.Runtime
                     }
                 }
             }
+            else if (axis.type == Axis.AxisType.Time)
+            {
+                if (ceilRate != 0)
+                {
+                    minValue = ChartHelper.GetMinCeilRate(minValue, ceilRate);
+                    maxValue = ChartHelper.GetMaxCeilRate(maxValue, ceilRate);
+                }
+            }
             else
             {
-                if (ceilRate == 0) ceilRate = axis.ceilRate;
                 switch (axis.minMaxType)
                 {
                     case Axis.AxisMinMaxType.Default:
-
                         if (minValue == 0 && maxValue == 0) { }
                         else if (minValue > 0 && maxValue > 0)
                         {
@@ -536,9 +542,9 @@ namespace XCharts.Runtime
         /// <param name="scaleWidth"></param>
         /// <param name="value"></param>
         /// <returns></returns>
-        public static float GetAxisValueLength(GridCoord grid, Axis axis, float scaleWidth, double value)
+        public static float GetAxisValueLength(GridCoord grid, Axis axis, float scaleWidth, double value, float gap = 0)
         {
-            return GetAxisPositionInternal(grid, axis, scaleWidth, value, false, true);
+            return GetAxisPositionInternal(grid, axis, scaleWidth, value, false, true, gap);
         }
 
         /// <summary>
@@ -572,10 +578,11 @@ namespace XCharts.Runtime
             }
         }
 
-        private static float GetAxisPositionInternal(GridCoord grid, Axis axis, float scaleWidth, double value, bool includeGridXY, bool realLength)
+        private static float GetAxisPositionInternal(GridCoord grid, Axis axis, float scaleWidth, double value, bool includeGridXY, bool realLength, float gap = 0)
         {
             var isY = axis is YAxis;
             var gridHeight = isY ? grid.context.height : grid.context.width;
+            gridHeight -= gap;
             var gridXY = isY ? grid.context.y : grid.context.x;
 
             if (axis.IsLog())