+++ Packages/XCharts/Runtime/Serie/Bar/BarHandler.cs
@@ -47,12 +47,14 @@ namespace XCharts.Runtime
             {
                 switch (label.position)
                 {
+                    case LabelStyle.Position.Start:
                     case LabelStyle.Position.Bottom:
                         var center = serieData.context.areaCenter;
                         var angle = serieData.context.halfAngle;
                         var radius = serieData.context.insideRadius;
                         return ChartHelper.GetPosition(center, angle, radius);
                     case LabelStyle.Position.Top:
+                    case LabelStyle.Position.End:
                         center = serieData.context.areaCenter;
                         angle = serieData.context.halfAngle;
                         radius = serieData.context.outsideRadius;
@@ -65,6 +67,7 @@ namespace XCharts.Runtime
             {
                 switch (label.position)
                 {
+                    case LabelStyle.Position.Start:
                     case LabelStyle.Position.Bottom:
                         var center = serieData.context.rect.center;
                         if (serie.context.isHorizontal)
@@ -73,11 +76,17 @@ namespace XCharts.Runtime
                             return new Vector3(center.x, center.y - serieData.context.rect.height / 2);
                     case LabelStyle.Position.Center:
                     case LabelStyle.Position.Inside:
+                    case LabelStyle.Position.Middle:
                         return serieData.context.rect.center;
                     default:
                         return serieData.context.position;
                 }
             }
+        }
+
+        public override Vector3 GetSerieDataTitlePosition(SerieData serieData, TitleStyle titleStyle)
+        {
+            return GetSerieDataLabelPosition(serieData, titleStyle);
         }
 
         private void UpdateSerieGridContext()
@@ -198,11 +207,13 @@ namespace XCharts.Runtime
             var dataChangeDuration = serie.animation.GetChangeDuration();
             var dataAddDuration = serie.animation.GetAdditionDuration();
             var interactDuration = serie.animation.GetInteractionDuration();
+            var exchangeDuration = serie.animation.GetExchangeDuration();
 
             var areaColor = ColorUtil.clearColor32;
             var areaToColor = ColorUtil.clearColor32;
             var interacting = false;
 
+            axis.context.scaleWidth = categoryWidth;
             serie.context.isHorizontal = isY;
             serie.containerIndex = m_SerieGrid.index;
             serie.containterInstanceId = m_SerieGrid.instanceId;
@@ -228,6 +239,7 @@ namespace XCharts.Runtime
                 var borderGap = relativedValue == 0 ? 0 : itemStyle.borderGap;
                 var borderGapAndWidth = borderWidth + borderGap;
                 var backgroundColor = itemStyle.backgroundColor;
+                var backgroundGap = itemStyle.backgroundGap;
 
                 if (!serieData.interact.TryGetColor(ref areaColor, ref areaToColor, ref interacting, interactDuration))
                 {
@@ -237,16 +249,21 @@ namespace XCharts.Runtime
 
                 var pX = 0f;
                 var pY = 0f;
-                UpdateXYPosition(m_SerieGrid, isY, axis, relativedAxis, i, categoryWidth, relativedCategoryWidth, barWidth, isStack, value, ref pX, ref pY);
-                var barHig = 0f;
+                UpdateXYPosition(m_SerieGrid, isY, axis, relativedAxis, i, categoryWidth, relativedCategoryWidth,
+                    barWidth, isStack, value, backgroundGap, ref pX, ref pY);
+                if (serie.useSortData)
+                {
+                    serieData.context.UpdateExchangePosition(ref pX, ref pY, exchangeDuration);
+                }
+                float barHig;
                 if (isPercentStack)
                 {
                     var valueTotal = chart.GetSerieSameStackTotalValue<Bar>(serie.stack, i, m_SerieGrid.index);
-                    barHig = valueTotal != 0 ? (float)(relativedValue / valueTotal * relativedAxisLength) : 0;
+                    barHig = valueTotal != 0 ? (float)(relativedValue / valueTotal * (relativedAxisLength - 2 * backgroundGap)) : 0;
                 }
                 else
                 {
-                    barHig = AxisHelper.GetAxisValueLength(m_SerieGrid, relativedAxis, relativedCategoryWidth, relativedValue);
+                    barHig = AxisHelper.GetAxisValueLength(m_SerieGrid, relativedAxis, relativedCategoryWidth, relativedValue, 2 * backgroundGap);
                 }
                 float currHig = AnimationStyleHelper.CheckDataAnimation(chart, serie, i, barHig);
                 Vector3 plb, plt, prt, prb, top;
@@ -257,8 +274,8 @@ namespace XCharts.Runtime
                 serieData.context.rect = Rect.MinMaxRect(plb.x + borderGapAndWidth, plb.y + borderGapAndWidth,
                     prt.x - borderGapAndWidth, prt.y - borderGapAndWidth);
                 serieData.context.backgroundRect = isY ?
-                    Rect.MinMaxRect(m_SerieGrid.context.x, plb.y, m_SerieGrid.context.x + relativedAxisLength, prt.y) :
-                    Rect.MinMaxRect(plb.x, m_SerieGrid.context.y, prb.x, m_SerieGrid.context.y + relativedAxisLength);
+                    Rect.MinMaxRect(m_SerieGrid.context.x, plb.y - backgroundGap, m_SerieGrid.context.x + relativedAxisLength, prt.y + backgroundGap) :
+                    Rect.MinMaxRect(plb.x - backgroundGap, m_SerieGrid.context.y, prb.x + backgroundGap, m_SerieGrid.context.y + relativedAxisLength);
 
                 if (!serie.clip || (serie.clip && m_SerieGrid.Contains(top)))
                 {
@@ -303,7 +320,7 @@ namespace XCharts.Runtime
 
         private void UpdateXYPosition(GridCoord grid, bool isY, Axis axis, Axis relativedAxis, int i,
             float categoryWidth, float relativedCategoryWidth, float barWidth, bool isStack,
-            double value, ref float pX, ref float pY)
+            double value, float backgroundGap, ref float pX, ref float pY)
         {
             if (isY)
             {
@@ -320,7 +337,7 @@ namespace XCharts.Runtime
                         pY = grid.context.y + valueLen - categoryWidth * 0.5f;
                     }
                 }
-                pX = AxisHelper.GetAxisValuePosition(grid, relativedAxis, relativedCategoryWidth, 0);
+                pX = AxisHelper.GetAxisValuePosition(grid, relativedAxis, relativedCategoryWidth, 0) + backgroundGap;
                 if (isStack)
                 {
                     for (int n = 0; n < m_StackSerieData.Count - 1; n++)
@@ -342,7 +359,7 @@ namespace XCharts.Runtime
                         pX = grid.context.x + valueLen - categoryWidth * 0.5f;
                     }
                 }
-                pY = AxisHelper.GetAxisValuePosition(grid, relativedAxis, relativedCategoryWidth, 0);
+                pY = AxisHelper.GetAxisValuePosition(grid, relativedAxis, relativedCategoryWidth, 0) + backgroundGap;
                 if (isStack)
                 {
                     for (int n = 0; n < m_StackSerieData.Count - 1; n++)